Add LTMs and STMs to score embed
